refactor: isolate remote bridge platform behavior
refs #3651
